pyproject.toml
# package build
# https://peps.python.org/pep-0517/
[build-system]
requires = ["setuptools>=65", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

# package basics
# https://peps.python.org/pep-0621/
[project]
name = "award-pynder"
description = "Award search and gathering"
keywords = []
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MPLv2" }
authors = [
  { email = "evamxb@uw.edu", name = "Eva Maxfield Brown" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dynamic = ["version"]
dependencies = [
  "pandas>=1",
  "python-dateutil>=2,<3",
  "requests>=2,<3",
  "tqdm>=4,<5",
]

[project.urls]
Homepage = "https://github.com/si2-urssi/award-pynder"
"Bug Tracker" = "https://github.com/si2-urssi/award-pynder/issues"
Documentation = "https://si2-urssi.github.io/award-pynder"
"User Support" = "https://github.com/si2-urssi/award-pynder/issues"

# extra dependencies
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
test = [
  "pytest>=8,<9",
]
lint = [
  "pre-commit>=2.20.0",
]
dev = [
  "ipython",
  "jupyterlab",
  "pyarrow>=10",
]

# entry points
# https://peps.python.org/pep-0621/#entry-points
# [project.entry-points."console_scripts"]


# build settings
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["*docs/*", "*tests/*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "py.typed"]

# tools
[tool.black]
line-length = 88
target-version = ["py312"]

# https://github.com/charliermarsh/ruff
[tool.ruff]
line-length = 88
target-version = "py311"
extend-select = [
    "E",      # style errors
    "F",      # flakes
    "D",      # pydocstyle
    "I001",   # isort
    "UP",     # pyupgrade
    "N",      # pep8-naming
    # "S",    # bandit
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "A001",   # flake8-builtins
    "RUF",    # ruff-specific rules
    "RUF100", # Unused noqa directive
]
extend-ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D103", # Missing docstring in public function
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D413", # Missing blank line after last section
    "D416", # Section name should end with a colon
]

[tool.ruff.per-file-ignores]
"**/notebooks/*" = ["E501"]
"**/tests/*.py" = ["D"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true